bug19549: Make canonicalization of / optional for schemes.
Add a keyword argument to parse-uri, :canonicalize-schemes, which
allows users to specify for which schemes canonicalization of the path
component "/" should be performed.

<documentation>
New keyword argument to `parse-uri`: :canonicalize-schemes. Specifies
for which the protocols parse-uri canonicalizes the empty path "/" to
"".

For example, if we were to parse-uri "http://franz.com/" and
"http://franz.com" with :canonicalize-schemes '(:http), uri= would
return t for the parsed uri.

Defaults to '(:http :https :ftp).
</documentation>