Implement rfe11850, rfe11851, rfe11852 (partial)
rfe11850: we canonicalize all these uri strings the same way:

       http://example.com
       http://example.com/
       http://example.com:/
       http://example.com:80/

as "http://www.franz.com".  RFC 3986 says:

  In general, a URI that uses the generic syntax for authority with an
  empty path should be normalized to a path of "/".

However, in the merging section, the RFC says this:

  If the base URI has a defined authority component and an empty
  path, then ...

This implies that an empty path exists and it wasn't normalized to be
"/".  And there's good reason for this: the merging rules would
be broken if we normalized path-less to be "/", according to the RFC's
own text and examples.

So, we take it that we should, according to the RFC, print the
trailing slash when the uri has a scheme and authority.

The problem is, this change would break existing programs.  So, we
introduce a new (undocumented, unexported) variable:

   net.uri::*render-include-slash-on-null-path*

This controls the printing of the final slash:
  When nil (the default), there is no change from current behavior.
  When non-nil, uri's with a null path and a scheme and host will print
      with a trailing slash.  Like this:

      (parse-uri "http://www.franz.com") => #<uri http://www.franz.com>
      (setq net.uri::*render-include-slash-on-null-path* t) => t
      (parse-uri "http://www.franz.com") => #<uri http://www.franz.com/>

Added tests for the above.

******************

rfe11851: remove canonicalization of null fragment

<release-notes>
Previously, we normalized a uri with a null fragment away such that it
was equivalent to a uri without a fragment.  That is, parsing the
following uri's produced the same uri:

  http://www.exampe.com/path#
  http://www.exampe.com/path

RFC 3986 seems pretty clear that an undefined fragment (the # was
missing) and a defined, though null, fragment should be handled
differently.  This change causes the null fragment to be preserved.
</release-notes>

Added tests for the above.

******************

rfe11852: leading . and .. elements are now removed on merge

<release-notes>
RFC 3986, unlike RFC 2396 (on which the original net.uri
implementation is based), is pretty clear about extra . and ..'s at
the beginning of the uri path: remove them.

Parsing does not remove extra . and ..'s, but merging now does.
</release-notes>

Added tests for the above.

******************

Also, add a copy of RFC 3986 to the repo.

Change-Id: I03f814d4195ef639fea79debe52e64dcb64963ec