bug25517/bug25523: uri fixes
bug25523: a typo in the new parser cause render-uri in upper case lisp
to insert garbage into the output.

bug25517: merge-uris gives the wrong answer in some cases.  All
fixed.  BIG CHANGE to parse-uri to make this work.  See below.

<release-note>
bug25517: parse-uri no longer canonicalizes away a path of "/"
</release-note>

<documentation>
bug25517: parse-uri no longer canonicalizes away a path of "/"

This change was made to allow for conformance to the RFCs.  It means
that

  (uri= (parse-uri "http://foo.com/") (parse-uri "http://foo.com"))

now returns NIL and not T as it did before.
</documentation>

Change-Id: Ied30cb6fbcbe71955567c57e4a6639ff91357cc7
Reviewed-on: https://gerrit.franz.com:9080/11778
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>