rfe15844: string-to-uri and uri-to-string
See documentation below.

Also included: parse-uri: minor fiddling with escape keyword and
usage.  Simplification.

<release-note>
rfe15844: string-to-uri and uri-to-string
</release-note>

<documentation>
rfe15844: string-to-uri and uri-to-string

parse-uri has been modified, for parsing the query of a URI, to not
decode #\+, #\= and #\& if they are percent encoded.  This means %2b,
for example, will not be converted to "+".  This is contrary to RFC
3986, but otherwise there is no way to pass these characters as field
values.

parse-uri/render-uri and string-to-uri/uri-to-string are pairs which
should be used together.  Mixing usage on the same URI object will
result in confusion.

string-to-uri string
  Parse STRING as a URI and either signal an error if it cannot be
  parsed or return the URI object.  This function differs from parse-uri
  in that the query is not decoded.  The knowledge of how to properly
  decode the query is outside the bounds of RFC 3986.

uri-to-string uri
  The inverse of string-to-uri, which returns the string
  representation of URI.
</documentation>

%Testing: t-uri.cl

Change-Id: I9d6014020b05b010a8a57692ff79fea43807a557
Reviewed-on: https://gerrit.franz.com:9080/11877
Tested-by: Kevin Layer <layer@franz.com>
Reviewed-by: Robert Rorschach <rfr@franz.com>
Reviewed-by: Duane Rettig <duane@franz.com>
Reviewed-by: John Foderaro <jkf@franz.com>
Reviewed-by: Martin Mikelsons <mm@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>