rfe15977: IRI support
Support IRI parsing without any decoding or encoding of UTF-8 encoded
characters.

<release-note>
rfe15977: add IRI support
</release-note>

<documentation>
rfe15977: add IRI support

IRI support has been added to the URI module.

Internationalized Resource Identifiers (IRIs) extend the syntax of
URIs to a much wider repertoire of characters.  RFC 3987 defines IRIs.
The main difference between URIs and IRIs is that IRIs can be composed
of characters from the Universal Character Set, the coded character
set defined by ISO/IEC 10646 [ISO10646] and the Unicode Standard.  In
contrast, the character set for URIs is ASCII.  The set of
non-reserved characters that compose IRIs is:

   ucschar        = %xA0-D7FF / %xF900-FDCF / %xFDF0-FFEF
                  / %x10000-1FFFD / %x20000-2FFFD / %x30000-3FFFD
                  / %x40000-4FFFD / %x50000-5FFFD / %x60000-6FFFD
                  / %x70000-7FFFD / %x80000-8FFFD / %x90000-9FFFD
                  / %xA0000-AFFFD / %xB0000-BFFFD / %xC0000-CFFFD
                  / %xD0000-DFFFD / %xE1000-EFFFD

The characters allowed in the query are:

   iprivate       = %xE000-F8FF / %xF0000-FFFFD / %x100000-10FFFD

The user-visible changes to the URI module are below.

* net.uri:iri                        [class]

  This is a subclass of net.uri:uri and objects returned by
  string-to-iri are instances of this class.

NOTE: because net.uri:iri is a subclass of net.uri:uri, uri-p will
      return T for instances of net.uri:iri.  However, from a
      specification standpoint, URIs are a subset of IRIs, which could
      be thought of as a subclass, of sorts.  We acknowledge this
      difference in CLOS and specification hierarchies.  The
      current design was chosen for better compatibility.

* iri-p                              [method]

  A predicate that returns T for instances of the class net.uri:iri,
  and nil for all other objects.

The following two functions are IRI analogs to string-to-uri and
uri-to-string.  The only difference in the new functions is the
characters allowed during parsing IRI strings.

* string-to-iri string               [function]

  Just like string-to-uri, this converts a string to an IRI.

* iri-to-string iri                  [function]

  Just like uri-to-string, this converts an IRI to a string.
</documentation>

Change-Id: If3acdaf74e1bbb76db923c6ef3fc04d59bbb2747
Reviewed-on: https://gerrit.franz.com:9080/12480
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>