rfe16179: final fix for patch loading
The previous version mistakenly assumed all previous versions were on
sys::*patches*, but they are not, so we need to check that the version
we find is >= to the version we want.

Also, move the patch loading form to after the in-package, so we do
not pollute the user package.

Lastly, changes for code clarity.

%Testing: manual

Change-Id: I56e7371796062bc4cefaf33dcefbd3e2e673b4aa
Reviewed-on: https://gerrit.franz.com:9080/13641
Tested-by: Kevin Layer <layer@franz.com>
Reviewed-by: Robert Rorschach <rfr@franz.com>
Reviewed-by: Duane Rettig <duane@franz.com>
Reviewed-by: Martin Mikelsons <mm@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>