bug26155: fix handling of %25 in the path
Previously, render-uri and uri-to-string both used uri-parsed-path
instead of uri-path to construct the string they printed/returned.
There was, seemingly, no reason for this.  It was certainly not
documented in the code.  Using uri-path allows the test suite to
pass.

Additional tests added to t-uri.cl for failing test case.

Change-Id: Ieece9a1603683c67305d0f4e02fd81dcf7cad363
Reviewed-on: https://gerrit.franz.com:9080/c/uri/+/16287
Tested-by: Kevin Layer <layer@franz.com>
Reviewed-by: Duane Rettig <duane@franz.com>
Reviewed-by: Robert Rorschach <rfr@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>
Reviewed-by: Martin Mikelsons <mm@franz.com>